requirements.txt
kafka-python==2.2.0
python-dateutil==2.8.2